Update settings.yml
To obtain the confusion matrix we first need to create the chow-liu tree by means of training data BoW representation and clustering centroid of the extracted SURF-128 features. Here we need to set the function in the setting file to TrainChowLiuTree to obtain this tree. Next, we need to GenerateFABMAPTestData and run the code using the obtained chow-liu tree in the previous step. Finally, we need to run the whole algorithm (by setting the FUNCTION variable to RunOpenFABMAP in both settings.yml and openFABMAPcli.cpp) using all the files that are created offline (Both vocabulary for the training data and Both BoW representation for the training and testing data).
It is mentionable that in the vocabulary generation step the labels are obtained from an autoencoder instead of  clustering algorithm.